requirements.txt
fastapi
uvicorn[standard]
python-dotenv
llama-index
llama-index-llms-openai
llama-index-embeddings-huggingface
llama-index-postprocessor-sentence-transformer-rerank
sentence-transformers
numpy
ragas
showdown-js
bootstrap
scikit-learn
torch
pandas
matplotlib
jupyter
nest_asyncio
psycopg2-binary
requests
beautifulsoup4
lxml
